Complete ability to copy previous microcycle
