+++ src/assets/stylesheets/components/_Exercise.scss
@@ -155,10 +155,4 @@
       background-color: $color-beige;
     }
   }
-
-  .copy-container {
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-  }
 }
