+++ src/scenes/dashboard/DailyWorkout.tsx
@@ -1,7 +1,12 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import Carousel from "../../components/Carousel";
 import useAxios from "../../hooks/useAxios";
 import { useDailyWorkout } from "../../contexts/DailyWorkoutContext";
+import Button from "../../components/Button";
+import { Microcycle } from "./Microcycle";
+import { postData } from "../../utils/api";
+import { useMicrocycles } from "../../contexts/MicrocyclesContext";
+import { IoCopy } from "react-icons/io5";
 
 interface DailyWorkoutProps {
   activeMicrocycle: number;
@@ -23,6 +28,9 @@ const DailyWorkout: React.FC<DailyWorkoutProps> = ({
   edittable,
 }) => {
   const { dailyWorkoutList, setDailyWorkoutList } = useDailyWorkout();
+  const [copyIndex, setCopyIndex] = useState<number | null>(null);
+
+  const { microcycles } = useMicrocycles();
 
   const { data, loading, fetchData } = useAxios<DailyWorkout[]>(
     [],
@@ -31,24 +39,55 @@ const DailyWorkout: React.FC<DailyWorkoutProps> = ({
     true
   );
 
-  const appendDailyWorkout = (newDailyWorkout: DailyWorkout) => {
-    setDailyWorkoutList([...dailyWorkoutList, newDailyWorkout]);
-  };
-
   useEffect(() => {
     fetchData();
+    setCopyIndex(findCopyId(microcycles, activeMicrocycle));
   }, [activeMicrocycle]);
 
   useEffect(() => {
     setDailyWorkoutList(data);
   }, [data]);
 
+  const appendDailyWorkout = (newDailyWorkout: DailyWorkout) => {
+    setDailyWorkoutList([...dailyWorkoutList, newDailyWorkout]);
+  };
+
+  const copyPreviousWeek = async () => {
+    const response = await postData(
+      `/admin/workout_programs/copy_previous_week`,
+      { previousMicrocycleId: copyIndex, newMicrocycleId: activeMicrocycle },
+      true
+    );
+    response && setDailyWorkoutList(response?.data.dailyWorkouts);
+  };
+
+  const findCopyId = (microcycles: Microcycle[], activeMicrocycle: number) => {
+    const targetIndex = microcycles.findIndex(
+      (microcycle) => microcycle.id === activeMicrocycle
+    );
+
+    if (targetIndex > 0) {
+      return microcycles[targetIndex - 1].id;
+    }
+
+    return null;
+  };
+
   if (loading) {
     return <div>Loading...</div>;
   }
 
   return (
-    <>
+    <div className="daily-workout-content">
+      {dailyWorkoutList.length === 0 && (
+        <div className="copy-container">
+          <Button size="small" onClick={copyPreviousWeek}>
+            <div className="copy-text">
+              Copy From Previous Workout <IoCopy />
+            </div>
+          </Button>
+        </div>
+      )}
       <Carousel
         revertCarouselReset={revertCarouselReset}
         resetCarousel={resetCarousel}
@@ -58,7 +97,7 @@ const DailyWorkout: React.FC<DailyWorkoutProps> = ({
         handleAdd={appendDailyWorkout}
         microcycleId={activeMicrocycle}
       />
-    </>
+    </div>
   );
 };
